Add Firebase configuration for Android
This commit adds the necessary Firebase configuration options for the Android platform to the `firebase_options.dart` file. The configuration includes the API key, app ID, messaging sender ID, project ID, and storage bucket. This configuration is generated by the FlutterFire CLI and is required to integrate Firebase services into the Flutter app.